Rename "uptime" command to "about" command.
